+++ packages/bootstrap-vue-next/src/components/BModal/BModal.vue
@@ -20,7 +20,7 @@
         tabindex="-1"
         v-bind="$attrs"
         :style="computedZIndex"
-        @click.self="hideFn('close-modal')"
+        @click.self="hideFn('backdrop')"
       >
         <div class="modal-dialog" :class="modalDialogClasses">
           <div v-if="lazyShowing" class="modal-content" :class="props.contentClass">
@@ -102,12 +102,14 @@
       </div>
     </Transition>
     <slot v-if="!props.hideBackdrop" name="backdrop" v-bind="sharedSlots">
-      <Transition @after-enter="fadeIn" @before-leave="fadeOut">
+      <Transition :appear="modelValue" @after-enter="fadeIn" @before-leave="fadeOut">
         <div
           v-if="modelValue"
           class="modal-backdrop"
+          :style="computedZIndexBackdrop"
           :class="{
             fade: !props.noFade,
+            show: props.noFade,
           }"
           @click="hideFn('backdrop')"
         />
@@ -196,7 +198,7 @@ const _props = withDefaults(defineProps<Omit<BModalProps, 'modelValue'>>(), {
   teleportTo: 'body',
   title: undefined,
   titleClass: undefined,
-  titleSrOnly: false,
+  titleVisuallyHidden: false,
   titleTag: 'h5',
   transProps: undefined,
 })
@@ -348,7 +350,7 @@ const footerClasses = computed(() => [props.footerClass, footerColorClasses.valu
 const titleClasses = computed(() => [
   props.titleClass,
   {
-    ['visually-hidden']: props.titleSrOnly,
+    ['visually-hidden']: props.titleVisuallyHidden,
   },
 ])
 const disableCancel = computed(() => props.cancelDisabled || props.busy)
@@ -482,16 +484,23 @@ watch(stackWithoutSelf, (newValue, oldValue) => {
 })
 
 const defaultModalDialogZIndex = 1056
-const computedZIndex = computed<CSSProperties>(() => ({
+const computedZIndexNumber = computed<number>(() =>
   // Make sure that newly opened modals have a higher z-index than currently active ones.
   // All active modals have a z-index of ('defaultZIndex' - 'stackSize' - 'positionInStack').
   //
   // This means inactive modals will already be higher than active ones when opened.
-  'z-index':
-    isActive.value || isLeaving.value
-      ? // Just for reference there is a single frame in which the modal is not active but still has a higher z-index than the active ones due to _when_ it calculates its position. It's a small visual effect
-        defaultModalDialogZIndex - ((activeModalCount?.value ?? 0) - (activePosition?.value ?? 0))
-      : defaultModalDialogZIndex,
+
+  isActive.value || isLeaving.value
+    ? // Just for reference there is a single frame in which the modal is not active but still has a higher z-index than the active ones due to _when_ it calculates its position. It's a small visual effect
+      defaultModalDialogZIndex -
+      ((activeModalCount?.value ?? 0) * 2 - (activePosition?.value ?? 0) * 2)
+    : defaultModalDialogZIndex
+)
+const computedZIndex = computed<CSSProperties>(() => ({
+  'z-index': computedZIndexNumber.value,
+}))
+const computedZIndexBackdrop = computed<CSSProperties>(() => ({
+  'z-index': computedZIndexNumber.value - 1,
 }))
 
 useEventListener(element, 'bv-toggle', () => {