+++ packages/bootstrap-vue-next/src/composables/useMouse.ts
@@ -1,0 +1,3 @@
+import {useMouse as _useMouse, createSharedComposable} from '@vueuse/core'
+
+export const useMouse = createSharedComposable(_useMouse)